requirements.txt
# AI/ML Core Dependencies
openai==1.7.2
numpy==1.24.3
pandas==2.0.3
pinecone-client==3.0.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
PyPDF2==3.0.1

# Additional utilities for robust implementation
requests==2.31.0
aiofiles==23.2.0
python-multipart==0.0.6